python/requirements.txt
bs4
google_cloud
pymongo
requests
textract
